+++ app/src/main/java/dev/tuandoan/expensetracker/ui/screen/home/HomeScreen.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -236,7 +235,6 @@ fun HomeScreen(
                         }
                     },
                     scrollBehavior = scrollBehavior,
-                    windowInsets = WindowInsets(0, 0, 0, 0),
                 )
             }
         },
@@ -254,7 +252,6 @@ fun HomeScreen(
             }
         },
         snackbarHost = { SnackbarHost(snackbarHostState) },
-        contentWindowInsets = WindowInsets(0, 0, 0, 0),
         modifier = modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
     ) { innerPadding ->
         Column(
@@ -943,7 +940,6 @@ private fun SearchTopBar(
             }
         },
         scrollBehavior = scrollBehavior,
-        windowInsets = WindowInsets(0, 0, 0, 0),
         modifier = modifier,
     )
 }