+++ app/src/main/java/dev/tuandoan/expensetracker/ui/screen/settings/SettingsScreen.kt
@@ -16,7 +16,6 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -176,11 +175,9 @@ fun SettingsScreen(
             TopAppBar(
                 title = { Text(stringResource(R.string.screen_title_settings)) },
                 scrollBehavior = scrollBehavior,
-                windowInsets = WindowInsets(0, 0, 0, 0),
             )
         },
         snackbarHost = { SnackbarHost(snackbarHostState) },
-        contentWindowInsets = WindowInsets(0, 0, 0, 0),
         modifier = modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
     ) { innerPadding ->
         Column(