+++ app/src/main/java/dev/tuandoan/expensetracker/ui/screen/summary/SummaryScreen.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -101,10 +100,8 @@ fun SummaryScreen(
             TopAppBar(
                 title = { Text(stringResource(R.string.screen_title_summary)) },
                 scrollBehavior = scrollBehavior,
-                windowInsets = WindowInsets(0, 0, 0, 0),
             )
         },
-        contentWindowInsets = WindowInsets(0, 0, 0, 0),
         modifier = modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
     ) { innerPadding ->
         Box(